move search_results policy filter to own filter, re #11477
