+++ arches/app/search/components/advanced_search.py
@@ -74,7 +74,7 @@ class AdvancedSearch(BaseSearchFilter):
         return search_query
 
     def append_dsl(self, search_query_object, **kwargs):
-        dsl = self.generate_dsl(search_query_object, kwargs)
+        dsl = self.generate_dsl(search_query_object, **kwargs)
         search_query_object["query"].add_query(dsl)
 
     def view_data(self):
