+++ arches/app/search/components/search_results.py
@@ -30,23 +30,13 @@ details = {
 
 
 class SearchResultsFilter(BaseSearchFilter):
-    def append_dsl(self, search_query_object, **kwargs):
+
+    def generate_dsl(self, search_query_object, **kwargs):
         permitted_nodegroups = kwargs.get("permitted_nodegroups")
         include_provisional = kwargs.get("include_provisional")
         nested_agg = NestedAgg(path="points", name="geo_aggs")
         nested_agg_filter = FiltersAgg(name="inner")
         geo_agg_filter = Bool()
-
-        try:
-            search_query_object["query"].dsl["query"]["bool"]["filter"][0]["terms"][
-                "graph_id"
-            ]  # check if resource_type filter is already applied
-        except (KeyError, IndexError):
-            resource_model_filter = Bool()
-            permitted_graphids = get_permitted_graphids(permitted_nodegroups)
-            terms = Terms(field="graph_id", terms=list(permitted_graphids))
-            resource_model_filter.filter(terms)
-            search_query_object["query"].add_query(resource_model_filter)
 
         if include_provisional is True:
             geo_agg_filter.filter(
